refactor: Move storage utility functions new module
